+++ example/app/basic/basic-example.component.ts
