+++ example/app/basic/basic-example.model.ts
