--- example/app/dynamic-form/dynamic-form.component.html
@@ -1,7 +1,0 @@
-<dynamic-form-basic-example></dynamic-form-basic-example>
-
-<dynamic-form-bootstrap-example></dynamic-form-bootstrap-example>
-
-<!--<dynamic-form-foundation-example></dynamic-form-foundation-example>-->
-
-<!--<dynamic-form-material-example></dynamic-form-material-example>-->
