+++ example/app/foundation/foundation-example.component.html
