+++ example/app/foundation/foundation-example.model.ts
