+++ example/app/main.ts
@@ -1,17 +1,17 @@
 import {bootstrap} from "@angular/platform-browser-dynamic";
 import {Title} from "@angular/platform-browser";
 import {disableDeprecatedForms, provideForms} from "@angular/forms";
-import {ROUTER_PROVIDERS} from "@angular/router-deprecated";
 import {HTTP_PROVIDERS} from "@angular/http";
 
 import {AppComponent} from "./app.component";
+import {APP_ROUTER_PROVIDERS} from "./app.routes";
 
 bootstrap(AppComponent, [
+    APP_ROUTER_PROVIDERS,
+    HTTP_PROVIDERS,
     Title,
     disableDeprecatedForms(),
-    provideForms(),
-    ROUTER_PROVIDERS,
-    HTTP_PROVIDERS
+    provideForms()
 
 ]).then(
