+++ example/index.html
@@ -15,6 +15,14 @@
             padding: 0;
         }
 
+        nav {
+            padding: 3rem;
+        }
+
+        nav a {
+            padding: 1rem;
+        }
+
         .content-padding {
             padding: 0 8rem;
         }
@@ -78,7 +86,7 @@
 
 </head>
 
-<base href="/">
+<base href=".">
 
 <body>
