+++ example/package.json
@@ -19,6 +19,7 @@
     "@angular/http": "^2.0.0-rc.3",
     "@angular/platform-browser": "^2.0.0-rc.3",
     "@angular/platform-browser-dynamic": "^2.0.0-rc.3",
+    "@angular/router": "^3.0.0-alpha.8",
     "@angular/router-deprecated": "^2.0.0-rc.2",
     "@angular2-material/button": "^2.0.0-alpha.5-2",
     "@angular2-material/checkbox": "^2.0.0-alpha.5-2",
