+++ example/systemjs.config.js
@@ -34,7 +34,7 @@
         "@angular/http",
         "@angular/platform-browser",
         "@angular/platform-browser-dynamic",
-        "@angular/router-deprecated",
+        "@angular/router",
         "@angular/testing",
         "@angular/upgrade"
     ];
